refactor(data): remove season module and simplify media model
This commit removes the `EdgeSeasonModel` and related code, as season data is not directly used.

It also simplifies the `EdgeMediaModel.Media.Kind` enum to a `String` and updates the `EdgeMediaEntity.kind` to use the `MediaType` enum.

The `EdgeImageMapper` now has a dedicated `onEmbedded` function for handling lists of `EdgeImageModel`.

The `EdgeMediaModelConverter` is updated to correctly map the `broadcast`, `chapters`, `volumes`, `moreInfo`, `publishedFrom`, and `publishedTo` fields.